moved MailMessage to its own package
