+++ schedule/schedule-appengine/src/main/java/com/jasify/schedule/appengine/meta/mail/MailMessageMeta.java
@@ -1,29 +1,29 @@
-package com.jasify.schedule.appengine.meta;
+package com.jasify.schedule.appengine.meta.mail;
 
 //@javax.annotation.Generated(value = { "slim3-gen", "@VERSION@" })
 /** */
-public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jasify.schedule.appengine.model.MailMessage> {
+public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jasify.schedule.appengine.model.mail.MailMessage> {
 
     /** */
-    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, com.google.appengine.api.datastore.Key> id = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, com.google.appengine.api.datastore.Key>(this, "__key__", "id", com.google.appengine.api.datastore.Key.class);
+    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, com.google.appengine.api.datastore.Key> id = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, com.google.appengine.api.datastore.Key>(this, "__key__", "id", com.google.appengine.api.datastore.Key.class);
 
     /** */
-    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, java.util.Date> created = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, java.util.Date>(this, "created", "created", java.util.Date.class);
+    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, java.util.Date> created = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, java.util.Date>(this, "created", "created", java.util.Date.class);
 
     /** */
-    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, java.util.Date> modified = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, java.util.Date>(this, "modified", "modified", java.util.Date.class);
+    public final org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, java.util.Date> modified = new org.slim3.datastore.CoreAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, java.util.Date>(this, "modified", "modified", java.util.Date.class);
 
     /** */
-    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage> pathInfo = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage>(this, "pathInfo", "pathInfo");
+    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage> pathInfo = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage>(this, "pathInfo", "pathInfo");
 
     /** */
-    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage> from = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage>(this, "from", "from");
+    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage> from = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage>(this, "from", "from");
 
     /** */
-    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage> subject = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.MailMessage>(this, "subject", "subject");
+    public final org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage> subject = new org.slim3.datastore.StringAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage>(this, "subject", "subject");
 
     /** */
-    public final org.slim3.datastore.UnindexedAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, com.google.appengine.api.datastore.Blob> messageData = new org.slim3.datastore.UnindexedAttributeMeta<com.jasify.schedule.appengine.model.MailMessage, com.google.appengine.api.datastore.Blob>(this, "messageData", "messageData", com.google.appengine.api.datastore.Blob.class);
+    public final org.slim3.datastore.UnindexedAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, com.google.appengine.api.datastore.Blob> messageData = new org.slim3.datastore.UnindexedAttributeMeta<com.jasify.schedule.appengine.model.mail.MailMessage, com.google.appengine.api.datastore.Blob>(this, "messageData", "messageData", com.google.appengine.api.datastore.Blob.class);
 
     private static final org.slim3.datastore.CreationDate slim3_createdAttributeListener = new org.slim3.datastore.CreationDate();
 
@@ -40,12 +40,12 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
 
     /** */
     public MailMessageMeta() {
-        super("Msg", com.jasify.schedule.appengine.model.MailMessage.class);
+        super("Msg", com.jasify.schedule.appengine.model.mail.MailMessage.class);
     }
 
     @Override
-    public com.jasify.schedule.appengine.model.MailMessage entityToModel(com.google.appengine.api.datastore.Entity entity) {
-        com.jasify.schedule.appengine.model.MailMessage model = new com.jasify.schedule.appengine.model.MailMessage();
+    public com.jasify.schedule.appengine.model.mail.MailMessage entityToModel(com.google.appengine.api.datastore.Entity entity) {
+        com.jasify.schedule.appengine.model.mail.MailMessage model = new com.jasify.schedule.appengine.model.mail.MailMessage();
         model.setId(entity.getKey());
         model.setCreated((java.util.Date) entity.getProperty("created"));
         model.setModified((java.util.Date) entity.getProperty("modified"));
@@ -58,7 +58,7 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
 
     @Override
     public com.google.appengine.api.datastore.Entity modelToEntity(java.lang.Object model) {
-        com.jasify.schedule.appengine.model.MailMessage m = (com.jasify.schedule.appengine.model.MailMessage) model;
+        com.jasify.schedule.appengine.model.mail.MailMessage m = (com.jasify.schedule.appengine.model.mail.MailMessage) model;
         com.google.appengine.api.datastore.Entity entity = null;
         if (m.getId() != null) {
             entity = new com.google.appengine.api.datastore.Entity(m.getId());
@@ -76,20 +76,20 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
 
     @Override
     protected com.google.appengine.api.datastore.Key getKey(Object model) {
-        com.jasify.schedule.appengine.model.MailMessage m = (com.jasify.schedule.appengine.model.MailMessage) model;
+        com.jasify.schedule.appengine.model.mail.MailMessage m = (com.jasify.schedule.appengine.model.mail.MailMessage) model;
         return m.getId();
     }
 
     @Override
     protected void setKey(Object model, com.google.appengine.api.datastore.Key key) {
         validateKey(key);
-        com.jasify.schedule.appengine.model.MailMessage m = (com.jasify.schedule.appengine.model.MailMessage) model;
+        com.jasify.schedule.appengine.model.mail.MailMessage m = (com.jasify.schedule.appengine.model.mail.MailMessage) model;
         m.setId(key);
     }
 
     @Override
     protected long getVersion(Object model) {
-        throw new IllegalStateException("The version property of the model(com.jasify.schedule.appengine.model.MailMessage) is not defined.");
+        throw new IllegalStateException("The version property of the model(com.jasify.schedule.appengine.model.mail.MailMessage) is not defined.");
     }
 
     @Override
@@ -102,7 +102,7 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
 
     @Override
     protected void prePut(Object model) {
-        com.jasify.schedule.appengine.model.MailMessage m = (com.jasify.schedule.appengine.model.MailMessage) model;
+        com.jasify.schedule.appengine.model.mail.MailMessage m = (com.jasify.schedule.appengine.model.mail.MailMessage) model;
         m.setCreated(slim3_createdAttributeListener.prePut(m.getCreated()));
         m.setModified(slim3_modifiedAttributeListener.prePut(m.getModified()));
     }
@@ -128,7 +128,7 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
 
     @Override
     protected void modelToJson(org.slim3.datastore.json.JsonWriter writer, java.lang.Object model, int maxDepth, int currentDepth) {
-        com.jasify.schedule.appengine.model.MailMessage m = (com.jasify.schedule.appengine.model.MailMessage) model;
+        com.jasify.schedule.appengine.model.mail.MailMessage m = (com.jasify.schedule.appengine.model.mail.MailMessage) model;
         writer.beginObject();
         org.slim3.datastore.json.Default encoder0 = new org.slim3.datastore.json.Default();
         if(m.getId() != null){
@@ -163,8 +163,8 @@ public final class MailMessageMeta extends org.slim3.datastore.ModelMeta<com.jas
     }
 
     @Override
-    protected com.jasify.schedule.appengine.model.MailMessage jsonToModel(org.slim3.datastore.json.JsonRootReader rootReader, int maxDepth, int currentDepth) {
-        com.jasify.schedule.appengine.model.MailMessage m = new com.jasify.schedule.appengine.model.MailMessage();
+    protected com.jasify.schedule.appengine.model.mail.MailMessage jsonToModel(org.slim3.datastore.json.JsonRootReader rootReader, int maxDepth, int currentDepth) {
+        com.jasify.schedule.appengine.model.mail.MailMessage m = new com.jasify.schedule.appengine.model.mail.MailMessage();
         org.slim3.datastore.json.JsonReader reader = null;
         org.slim3.datastore.json.Default decoder0 = new org.slim3.datastore.json.Default();
         reader = rootReader.newObjectReader("id");
