+++ schedule/schedule-appengine/src/main/java/com/jasify/schedule/appengine/util/KeyUtil.java
@@ -4,7 +4,7 @@ import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableBiMap;
-import com.jasify.schedule.appengine.meta.MailMessageMeta;
+import com.jasify.schedule.appengine.meta.mail.MailMessageMeta;
 import com.jasify.schedule.appengine.meta.activity.*;
 import com.jasify.schedule.appengine.meta.application.ApplicationMeta;
 import com.jasify.schedule.appengine.meta.application.ApplicationPropertyMeta;
